requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
annotated-types=0.7.0=pypi_0
anyio=4.6.2.post1=pypi_0
bcrypt=3.2.2=pypi_0
blas=1.0=openblas
blinker=1.6.2=py310hca03da5_0
bottleneck=1.3.7=py310hbda83bc_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.9.24=hca03da5_0
cffi=1.17.1=pypi_0
click=8.1.7=py310hca03da5_0
ecdsa=0.19.0=pypi_0
exceptiongroup=1.2.0=py310hca03da5_0
fastapi=0.115.2=pypi_0
flask=3.0.3=py310hca03da5_0
h11=0.14.0=pypi_0
idna=3.10=pypi_0
iniconfig=1.1.1=pyhd3eb1b0_0
itsdangerous=2.2.0=py310hca03da5_0
jinja2=3.1.4=py310hca03da5_0
joblib=1.4.2=py310hca03da5_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libopenblas=0.3.21=h269037a_0
llvm-openmp=14.0.6=hc6e5704_0
markupsafe=2.1.3=py310h80987f9_0
ncurses=6.4=h313beb8_0
numexpr=2.8.7=py310hecc3335_0
numpy=1.26.4=py310h3b2db8e_0
numpy-base=1.26.4=py310ha9811e2_0
openssl=1.1.1w=h1a28f6b_0
packaging=24.1=py310hca03da5_0
pandas=2.2.2=py310h313beb8_0
passlib=1.7.4=pypi_0
pip=24.2=py310hca03da5_0
pluggy=1.0.0=py310hca03da5_1
pyasn1=0.6.1=pypi_0
pycparser=2.22=pypi_0
pydantic=2.9.2=pypi_0
pydantic-core=2.23.4=pypi_0
pytest=7.4.4=py310hca03da5_0
python=3.10.4=hbdb9e5c_0
python-dateutil=2.9.0post0=py310hca03da5_2
python-dotenv=1.0.1=pypi_0
python-jose=3.3.0=pypi_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2024.1=py310hca03da5_0
readline=8.2=h1a28f6b_0
rsa=4.9=pypi_0
scikit-learn=1.5.1=py310h46d7db6_0
scipy=1.12.0=py310h20cbe94_0
setuptools=75.1.0=py310hca03da5_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h80987f9_0
starlette=0.40.0=pypi_0
threadpoolctl=3.5.0=py310h33ce5c2_0
tk=8.6.14=h6ba3021_0
tomli=2.0.1=py310hca03da5_0
typing-extensions=4.12.2=pypi_0
tzdata=2024b=h04d1e81_0
uvicorn=0.32.0=pypi_0
werkzeug=3.0.3=py310hca03da5_0
wheel=0.44.0=py310hca03da5_0
xz=5.4.6=h80987f9_1
zlib=1.2.13=h18a0788_1
